Fix for arm64 generator
