+++ .gitignore
@@ -1,0 +1,2 @@
+.idea
+asm2plan9s
